requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==2.2.8
django-tables2-bootstrap4==2.0.0b5
djangorestframework==3.12.4
idna==2.10
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.3
